[SQUASH] overlay: SimpleDeviceConfig: Compact cached app heaps in the background & enable cached app freezer
device_config: Compact cached app heaps in the background

Background ART heap compaction for cached apps has been recommended for
Android Go devices for a while, and Google has enabled it for the Pixel
5 on RQ1A.201205.010. This feature should help fit more cached apps in
memory.

device_config: Allow more cached apps in the background

Google has provisioned this on the Pixel 5 (RQ1A.201205.010) to utilize
the 8 GB of RAM better, and I have also observed that a lightweight
selection of apps was hitting the cached app limit before exhausting RAM
on another device with 8 GB of RAM.

This shouldn't cause adverse effects on devices with less RAM (e.g. 4
GB) as the low memory killer should kick in long before this limit on
such devices.

device_config: Enable cached app freezer on supported devices

I've been daily driving this on the Pixel 5 for months with no issues
(since qpr1), and qpr2 improved it more. This should be safe to enable
for all redbull users now.

Signed-off-by: Omkar Chandorkar <gotenksIN@aosip.dev>
Signed-off-by: jhonboy121 <alfredmathew05@gmail.com>